feat:js add a confirm to ask sure from user to delete

+++ resources/views/components/todo/item/delete.blade.php
@@ -1,6 +1,6 @@
 @aware(['deletePath'])
 <div class="delete">
-    <a href="{{$deletePath}}">
+    <a href="{{$deletePath}}" onclick="return confirm('Are you sure to delete?')">
         <h2>
             Delete
         </h2>
